Fixes #5 - Removing IntegerHelper and moving stuff to NumberHelper
